Bug fix: Correct ping logic
For every connection, we have the 'heartbeat' timeout. This is a 60
second timeout that allows us to know whether or not a connection is
alive or has gone. The idea is that whenever a client talks *to* us, we
extend the counter by 60 seconds.

The previous code had a bug that meant whenever there was any activity,
we would reset the timer - activity that could have been caused by us
sending a message to the client.

Instead, we only extend the timer if we receive a message.